+++ src/main/java/com/application/close/modtcp/repo/TcpDataRepo.java
@@ -2,8 +2,8 @@ package com.application.close.modtcp.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.application.close.modtcp.entity.ModTcpData;
+import com.application.close.modtcp.entity.TcpData;
 
-public interface TcpDataRepo extends JpaRepository<ModTcpData, Integer> {
+public interface TcpDataRepo extends JpaRepository<TcpData, Integer> {
 
 }