+++ src/main/java/com/application/close/modtcp/service/ModService.java
@@ -1,0 +1,11 @@
+package com.application.close.modtcp.service;
+
+public interface ModService {
+
+	String connectToSlaveDevice(Integer tcpDataId);
+
+	String reconnect(Integer tcpDataId);
+
+	String disconnect(Integer tcpDataId);
+
+}